requirements.txt
Flask
Pillow
pyzbar
python-barcode
